Refactor useResourceLabel into useGetResourceLabel
